Clear the small warning categories
Thirteen signed/unsigned comparisons, all analysed and all benign: index
variables clamped to >= 0 before being compared against a COUNTOF, and one
INT64 price against an unsigned balance where the price cannot be
negative. Made explicit rather than left for the next reader to re-derive.

Four comparisons that can never be true, on types that are unsigned:
MULTI_TYPE and HUE_TYPE have no negative case, so the "< 0" halves of
those guards were dead.

Plus an unused local, a mixed enum/int conditional, and a class that
declared operator= without the matching copy constructor.

Visible warnings are now 82 on both architectures, from 659 when they were
first turned on. What remains is 59 implicit fallthroughs, 7 array-bounds
warnings that all come from one line of a system header, 7
address-of-packed-member on the packed network structs (by design), 3
"this == NULL" comparisons left deliberately, 2 in the vendored regex
header, and 4 dead comparisons whose types would have to change.

+++ src/common/CGrayInst.cpp
@@ -281,9 +281,9 @@ VERFILE_TYPE CGrayInstall::OpenFiles( unsigned long long dwMask )
 											unsigned long long qwHash = ((INT64)dwHashHi << 32) + dwHashLo;
 											m_Maps[index].Seek( sizeof(DWORD)+sizeof(WORD), SEEK_CUR );
 					
-											for (int x = 0; x < dwLoop; x++)
+											for (DWORD x = 0; x < dwLoop; x++)
 											{
-												sprintf(z, "build/map%dlegacymul/%.8d.dat", index, x);
+												sprintf(z, "build/map%dlegacymul/%.8d.dat", index, static_cast<int>(x));
 												if (HashFileName(z) == qwHash)
 												{
 													pMapAddress.dwFirstBlock = x*4096;
